Move stack related functions to util.go
