+++ notice.go
@@ -6,11 +6,8 @@ import (
 	"os"
 	"path/filepath"
 	"runtime"
-	"strconv"
 	"strings"
 	"sync"
-
-	"github.com/pkg/errors"
 )
 
 var defaultContextOnce sync.Once
@@ -64,46 +61,6 @@ func (n *Notice) String() string {
 	}
 	e := n.Errors[0]
 	return fmt.Sprintf("Notice<%s: %s>", e.Type, e.Message)
-}
-
-// stackTraces returns the stackTrace of an error.
-// It is part of the errors package public interface.
-type stackTracer interface {
-	StackTrace() errors.StackTrace
-}
-
-// getStack returns the stacktrace associated with e. If e is an
-// error from the errors package its stacktrace is extracted, otherwise
-// the current stacktrace is collected end returned.
-func getStack(e interface{}, depth int) []StackFrame {
-	if err, ok := e.(stackTracer); ok {
-		return stackFromErrorWithStackTrace(err)
-	}
-
-	return stack(depth)
-}
-
-// stackFromErrorWithStackTrace extracts the stacktrace from e.
-func stackFromErrorWithStackTrace(e stackTracer) []StackFrame {
-	var frames []StackFrame
-	for _, f := range e.StackTrace() {
-		line, _ := strconv.ParseInt(fmt.Sprintf("%d", f), 10, 64)
-		sf := StackFrame{
-			Func: fmt.Sprintf("%n", f),
-			File: fmt.Sprintf("%s", f),
-			Line: int(line),
-		}
-		frames = append(frames, sf)
-	}
-	return frames
-}
-
-// getTypeName returns the type name of e.
-func getTypeName(e interface{}) string {
-	if err, ok := e.(error); ok {
-		e = errors.Cause(err)
-	}
-	return fmt.Sprintf("%T", e)
 }
 
 func NewNotice(e interface{}, req *http.Request, depth int) *Notice {
